requirements.txt
# Core runtime (Python 3.10+; tested on 3.12+)
google-generativeai>=0.8.0
chromadb>=0.5.23
sentence-transformers>=3.3.0
python-dotenv>=1.0.1
pydantic>=2.10.0
rich>=13.9.0
tabulate>=0.9.0

# Dev
pytest>=8.3.0
